Add files via upload
speech to text recognition file. The Audio file is converted to the valid .wav format with a valid Riff header using pydub. It then uses the speech_Recognition module to give the audio transcription. this output is fed to the NLP model which does sentiment analysis and classifies the audio as danger or safe.